Moved SkillType to its own file to prepare for splitting seat data and seat function
