+++ source/entities/Creature.cpp
@@ -32,6 +32,7 @@
 #include "entities/Weapon.h"
 #include "game/Player.h"
 #include "game/Skill.h"
+#include "game/SkillType.h"
 #include "game/Seat.h"
 #include "gamemap/GameMap.h"
 #include "gamemap/Pathfinding.h"
