+++ source/network/ODClient.cpp
@@ -25,8 +25,9 @@
 #include "entities/Tile.h"
 #include "entities/Weapon.h"
 #include "game/Player.h"
-#include "game/Skill.h"
 #include "game/Seat.h"
+#include "game/Skill.h"
+#include "game/SkillType.h"
 #include "gamemap/GameMap.h"
 #include "modes/GameMode.h"
 #include "modes/MenuModeConfigureSeats.h"
