+++ source/network/ODServer.cpp
@@ -25,6 +25,7 @@
 #include "game/Player.h"
 #include "game/Skill.h"
 #include "game/SkillManager.h"
+#include "game/SkillType.h"
 #include "game/Seat.h"
 #include "gamemap/GameMap.h"
 #include "gamemap/MapLoader.h"
